pyproject.toml
[project]
name = "speech-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-whisper==20250625",
    "pillow==12.0.0",
    "pyperclip==1.11.0",
    "pystray==0.19.5",
    "sounddevice==0.5.3",
]

[project.gui-scripts]
speech-to-text = "main:main"

[dependency-groups]
dev = [
    "pyinstaller==6.17.0",
]
